Data.Currency.Convert

Description

This module allows values of currency to be converted from one currency to another using exchange rates from various sources.

The basic operation for this module is as follows:

First, get a conversion function (with getDefaultConverter or getConverter):

>>> Converter convert <- getDefaultConverter

Then, construct some value of currency using the convenience functions, and convert it using the convert function you just got:

>>> convert (usd 100) :: EUR
90.0 eur

The result type of convert defines what currency the value will be converted into.

Documentation

newtype Converter

Converter is a newtype wrapper around the type of conversion functions to avoid ImpredicativeTypes in getConverter

Constructors

Converter (forall a b. (KnownSymbol a, KnownSymbol b) => Currency a -> Currency b)

getConverter :: RateProvider -> IO Converter

getConverter takes a rate provider and returns a Converter in the IO monad. It is used like this:

>>> Converter convert <- getConverter provider

convert can then be used to convert between currencies:

>>> convert (usd 100) :: GBP
75.0 gbp

getDefaultConverter :: IO Converter

getDefaultConverter is like getConverter except no provider needs to be specified (defaultProvider is used).

unsafeCoerceCurrency :: Currency a -> Currency b

unsafeCoerceCurrency can be used to convert between currencies directly without using exchange rates. Only use this if you know the resulting currency is supported by your RateProvider

data Currency s

Currency is a wrapper around Double that has a phantom symbol type, allowing different currencies to be distinguished using type level literals.

To extract the value from a Currency value, use toRational from Real.

data RateProvider

RateProvider is used by getConverter to provide a conversion rate dictionary to the resulting conversion function. RateProvider has two fields, the name (:: String), and the action (:: IO RateDict). The name is used in error reporting, and the action is used to provide conversion rates when needed.

There are several prebuilt RateProviders - fixerIOProvider, backupProvider, localProvider, dictProvider and defaultProvider.

Constructors

RateProvider String (IO RateDict)

newtype RateDict

RateDict is the type provided by a RateProvider. It is a mapping from currency codes to currency-to-euro conversion rates.

Constructors

RateDict [(String, Double)]

defaultProvider :: RateProvider

This RateProvider is used by getDefaultConverter it first tries fixer.io and then resorts to the backup dictionary.

backupProvider :: RateProvider

This RateProvider has a hardcoded dictionary.

dictProvider :: String -> [(String, Double)] -> RateProvider

This RateProvider allows a dictionary to be provided.

localProvider :: FilePath -> RateProvider

This RateProvider takes a file path and returns a dictionary of rates from the file. The file must be in the form

    <name1> <rate1>
    <name2> <rate2>
    ...
   

fixerIOProvider :: RateProvider

This RateProvider uses the fixer.io API to return exchange rates updated daily by the European Central Bank.

(<|-|>) :: RateProvider -> RateProvider -> RateProvider infixr 9

'(<|-|>)' combines two RateProviders to create a third. This provider tries to return the first providers result, but if an exception is raised, it returns the second providers result.

aud :: Double -> AUD

The Australian Dollar

bgn :: Double -> BGN

The Bulgarian Lev

brl :: Double -> BRL

The Brazillian Real

cad :: Double -> CAD

The Canadian Dollar

chf :: Double -> CHF

The Swiss Franc

cny :: Double -> CNY

The Yuan

czk :: Double -> CZK

The Czech Karuna

dkk :: Double -> DKK

The Danish Krone

gbp :: Double -> GBP

The Pound Sterling

hkd :: Double -> HKD

The Hong Kong Dollar

hrk :: Double -> HRK

The Croatian Kuna

huf :: Double -> HUF

The Forint

idr :: Double -> IDR

The Rupiah

ils :: Double -> ILS

The New Israeli Sheqel

inr :: Double -> INR

The Indian Rupee

jpy :: Double -> JPY

The Yen

krw :: Double -> KRW

The Won

mxn :: Double -> MXN

The Mexican Peso

myr :: Double -> MYR

The Malaysian Ringgit

nok :: Double -> NOK

The Norwegian Krone

nzd :: Double -> NZD

The New Zealand Dollar

php :: Double -> PHP

The Phillipine Peso

pln :: Double -> PLN

The Zloty

ron :: Double -> RON

The Romanian Leu

rub :: Double -> RUB

The Russian Ruble

sek :: Double -> SEK

The Swedish Krona

sgd :: Double -> SGD

The Singapore Dollar

thb :: Double -> THB

The Thai Baht

try :: Double -> TRY

The Turkish Lira

usd :: Double -> USD

The United States of America Dollar

zar :: Double -> ZAR

The Rand

eur :: Double -> EUR

The Euro

type AUD = Currency "aud"

The Australian Dollar

type BGN = Currency "bgn"

The Bulgarian Lev

type BRL = Currency "brl"

The Brazillian Real

type CAD = Currency "cad"

The Canadian Dollar

type CHF = Currency "chf"

The Swiss Franc

type CNY = Currency "cny"

The Yuan

type CZK = Currency "czk"

The Czech Karuna

type DKK = Currency "dkk"

The Danish Krone

type GBP = Currency "gbp"

The Pound Sterling

type HKD = Currency "hkd"

The Hong Kong Dollar

type HRK = Currency "hrk"

The Croatian Kuna

type HUF = Currency "huf"

The Forint

type IDR = Currency "idr"

The Rupiah

type ILS = Currency "ils"

The New Israeli Sheqel

type INR = Currency "inr"

The Indian Rupee

type JPY = Currency "jpy"

The Yen

type KRW = Currency "krw"

The Won

type MXN = Currency "mxn"

The Mexican Peso

type MYR = Currency "myr"

The Malaysian Ringgit

type NOK = Currency "nok"

The Norwegian Krone

type NZD = Currency "nzd"

The New Zealand Dollar

type PHP = Currency "php"

The Phillipine Peso

type PLN = Currency "pln"

The Zloty

type RON = Currency "ron"

The Romanian Leu

type RUB = Currency "rub"

The Russian Ruble

type SEK = Currency "sek"

The Swedish Krona

type SGD = Currency "sgd"

The Singapore Dollar

type THB = Currency "thb"

The Thai Baht

type TRY = Currency "try"

The Turkish Lira

type USD = Currency "usd"

The United States of America Dollar

type ZAR = Currency "zar"

The Rand

type EUR = Currency "eur"

The Euro